Fixes move selector and move hue bug. (#18)
Moving the color and hue selectors uses a clamp function that ensures
the coordinates are bounded within the dimensions of the color
selector elements coordinates. Unfortunately it is supplied a value
that expects pixels but when using a unit of a different type
(vw, vh, %) it takes the unit value as the pixel value. eg. setting
width to 50% sets the saturationWidth to 50px. This means that on a
viewport with 500px width if the color picker is set to 50% the color
picker will show as being 250px wide but only 50px of that width is
clickable. The same bug is present for height and the hue selector
height.

Getting the width and height from the dom elements directly fixes
this bug.